Add models and textures, update references
This stuff wasn't added before because it was in the Debug folder,
which is listed in .gitignore.  Further, .obj files are ALSO
listed in .gitignore.  I pulled out the necessary resource files
into a new Resources folder and removed .obj from the .gitignore
file, and updated the references in the code to this new location.